Renamed components that have been fully ported over to es2015
